+++ frontend/src/main.jsx
@@ -2,6 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App.jsx";
+import ErrorBoundary from "./components/ErrorBoundary";
 import AuthProvider from "./context/AuthContext";
 import DataProvider from "./context/DataContext";
 import CrmProvider from "./context/CrmContext";
@@ -11,20 +12,22 @@ import UploadProvider from "./context/UploadContext";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
-    <BrowserRouter>
-      <AuthProvider>
-        <DataProvider>
-          <CrmProvider>
-            <TeamProvider>
-              <PricingProvider>
-                <UploadProvider>
-                  <App />
-                </UploadProvider>
-              </PricingProvider>
-            </TeamProvider>
-          </CrmProvider>
-        </DataProvider>
-      </AuthProvider>
-    </BrowserRouter>
+    <ErrorBoundary>
+      <BrowserRouter>
+        <AuthProvider>
+          <DataProvider>
+            <CrmProvider>
+              <TeamProvider>
+                <PricingProvider>
+                  <UploadProvider>
+                    <App />
+                  </UploadProvider>
+                </PricingProvider>
+              </TeamProvider>
+            </CrmProvider>
+          </DataProvider>
+        </AuthProvider>
+      </BrowserRouter>
+    </ErrorBoundary>
   </StrictMode>
 );